Added validation of -f and -b flags
It is an error to attempt to use the `-f` and `-b` flags together. Note
that since both of these flags take string arguments, `qris -f -b` and
`qris -b -f` are both currently accepted as flags with their arguments
being the second flag. I'm not too concerned with this right now, but
maybe later....