Added ability to save config settings
The `-dir` flag now saves the new working directory path to a config
file located in a system-specified location. A new `-config` flag can be
used to display the path to the configuration file. This path will be
shown even if no configuration file has been created. A configuration
file is automatically created when a new working directory location is
saved, and the configuration file is overwritten whenever a new location
is saved.