Added note parsing functionality
Lines ending with "jmr" are considered notes.

If a note is found before any quotes have been read, the note is assumed
to belong to the citation, and a citation note is recorded for that
citation. Exported in the CN tag.

If a note is found after quotes have been read, the note is assumed to
belong to the most recent quote, and a citation note is recorded for
that quote. Exported in the CY tag.