Added comments about quote-end parsing
Jack has some problems with a few quotes not being identified as quotes.
These were from older files, and in those cases there was no tab
delimiter between the body of the quote and the page number indicating
the end of the quote.

This is not a problem that can be solved by a better regex, as far as I
can tell, since without the tab delimiter the last page number of a line
must be used to identify the end of a quote. The problem with this is
that there may be annotations and comments following the page number
associated with the quote (the supplementary field of the quote), and
this information might (and in some instances does) include another page
number.

I can imagine using the presence of a "w^" marker to ascertain the
correct page number, i.e., the page number is the last page number of a
quote, unless there is a "w^" marker present, in which case the page
number is the one immediately preceding the "w^" marker.

I think that this could work, if Jack's data is consistent in that
regard, but it seems fragile, and for now I will just suggest manually
inserting tab delimiters when necessary.