Fixed bug in parseQuote
In some instances the input quotes have page number indicators with no
following page numbers, e.g., "pp.  ".

The `quoteEnd` regex does not look for numbers following a page
indicator, but the `quotePage` regex must look for numbers to split the
end of a quote into a page number and a supplementary field.

When this occurs, and empty slice of string indices was returned
representing the two parts of the quote end. The code then attempted to
index into this empty slice, causing a panic.

Since `FindStringIndex` returns `nil` when there is no match, I added a
test to avoid this situation. When there is no match for a page number
in a quote end, the page value is reported as the string "UNKNOWN".