Added comments about tagging changes
Jack mentioned that he doesn't feel like the teaser is needed anymore.
The T3 tag can be entirely removed, and the T1 tag should now be
associated with the quote body `q.Body`. This means that `abst` and
`abstSize` can be remove, also.

I will wait to implement this until Jack provides a list of similar
changes in the near future.