Fixed bug in supplementary field parsing
Supplementary fields ending with "jmr" were picked up as notes and
attached to the previous quote. This was fixed by ensuring that a line
parsed as a note is not also a quote (with a tab-delimited page number
ending).

I wondered if I should add supplementary fields with note markers to the
`Note` field of a quote instead of the supplementary field. I decided
that it would better to leave things as is and let such notes be placed
in the supplementary field. Jack agreed; this allows a note following a
quote which happens to end with a note to record both notes, one in the
supplementary field and one in the note field.