Fixed bug in file parsing loop
When a directory was found in the folder containing files to be parsed,
the directory was treated as a file to be parsed, causing a panic when
there were no lines to index. This highlights some other fragility in
indexing into slices without validation, but for now the fix was to
check files for a ".txt" ending. Any file with a ".txt" ending in the
specified directory will be parsed. It is possible that a directory will
have such an ending; the program would attempt to parse the directory
and panic in that case. I will address this and other fragilities soon.

I also changed the endings on DISCARD files to "_DISCARD.log". The
upshot of this fix combined with the new file ending is that now only
.txt files are processed, so you can run the parsing routine a second
time on the same files without deleting the first output.