Created GetWorkDir function in qris.go
Moved functionality for getting a working directory from the
configuration file into the GetWorkDir function in qris.go.

Updated error printing to go to Stderr.

Moved the call to qris.GetWorkDir to after the code in main.go that
handles the -d flag to set a new working directory. This way the
program prints the updated working directory when it informs the user
about the working directory.