Fixed bug in single-file processing.
I managed to introduce a bug in handling the single-file parsing case.
The argument to `-f` was not being added to `dataList`. I was cleaning
up the logic here so that a bad warning about always using one of the
flags `-f` or `-b` would no longer be issued, and when testing this on a
single file I found the bug.