Added quoteEndAlt regexp
Added `quoteEndAlt` regexp to detect quotes ending with
non-tab-delimited page numbers. This regexp looks for page numbers
delimited by three or more spaces.

I also extended the special case search range to 30 characters. This is
a final check to determine if there is a page number present in the last
30 characters of a line. It is possible that this larger search range
could pick up page numbers earlier in the body of a quote that are not
the actual page number for the quote, but this has not happened in any
of my test files.