introduced Source data type
In qris.go:
-----------
Introduced `Source` type to encapsulate a citation
and associated quotes.

Modified `WriteQuotes` to use the new `Source` field
of a `ParsedFile`.

In parse.go:
------------
Modifed `ParsedFile` to use a `Source` field instead of separate
`Citation` and `Quotes` fields.

Modified `newParsedFile` to accept a `Source` argument instead of
separate `Citation` and `[]Quote` arguments.

Modified `ParseFile` to call `newParsedFile` with a `Source` argument.