fixed bug with single/multi-line quotes
The `ParseFile` function wasn't handling the case where the first line
of a multi-line quote ends in a (tab-delimited) page number. This could
happen if a user accidentally used the multi-line tag while intending a
single-line quote, or it could happen after an edit that changed the
body of the quote so that it fits on one line. This could result in
lost and/or spurious quote information.

Fixed by first checking whether a multi-line quote line is actually a
single-line quote; if so the line is cleaned up and processed as a
single-line quote, otherwise the multi-line state is initiated.