added -datestamp flag
Added a `-datestamp` flag which defaults to `true`, allowing
selective omission of the AD datestamp field from output files.

This solved a problem with testing: the _EXPECT.ris files had
datestamp fields, but those would not agree with the corresponding
fields created during testing which might occur on different dates.

This argument is passed to `WriteResults`, which in turn passes it to
`WriteQuotes`, where it controls the writing of datestamps to output.