improved TestParseQuote
Created a `testCase` type and moved test cases into a slice of `testCases`.

This makes it much easier to see the input values and their expected output
values because they are together, and also makes writing future tests
easier and less error-prone.

This change also allowed considerable simplification of the testing code.