fixed newline inconsistency in _DISCARD files
The _DISCARD output from .docx files contained extra newlines
between every line when compared with the same output from a
comparable .txt file.

This turned out to be because code in `DocxToLines` was adding an extra
newline after assembling a run from the .docx file XML source. There is
no need for this extra newline; it was a mistake, though it didn't cause
any problems in the main output of the program.

The fix was simply to remove the line which added the newline. After the
change _DISCARD outputs matched.

The program still passed all tests after this change.

I also updated some related comments in `qris.go`.