modified multiline quote output format
Multiline quotes no longer preserve leading whitespace or newlines.

Each line of a multiline quote is now given its own T1 tag and line
in the RIS output file.

Note that testing seems to be failing in `TestWriteResults`.

I don't yet understand the problem: when I rename the output file to
_EXPECT.ris, the tests fail. This makes no sense, and running `diff`
to compare the _PARSED.ris file with the _EXPECT.ris file shows no
differences. But when I check the length of the byte slices obtained
by reading these files they don't match!

I ran these tests on a Debian laptop, but until today I did all
development on a Windows laptop (but in Msys2). Maybe this is some
encoding issue?

Committing for now since I think that the program is working correctly.