added -lend line ending flag
Added a flag for setting output line endings.

This allows the user to set line endings to 'lf', 'crlf', or
'platform' (the user's platform line ending: this is the default).

Changed `lineEnding` to `LineEnding` in the `qris` package so that
this variable can be modified in the `main` package.

The main purpose of this change is to allow tests to be run using LF
endings. The repository uses Unix LF line endings, but using LF
endings in a Windows development environment leads to testing failures
if the line endings of output files are platform-specific (CRLF). This
change allows tests to specify a consistent line ending style so that
_EXPECT files match output files.