fixed bug with non-breaking hyphens in .docx input
MS Word encodes non-breaking hyphens as XML elements in the .docx file.

These non-breaking hyphens were being ignored when the text content was
extracted from a .docx file. The problem manifested as a page parsing
problem when the page range "161-180" at the end of a citation was fetched
as the number "161180". This exposed a problem with the citation page number
regexp (one of the first written for this project).

I rewrote the regexp to be more particular, and added another regexp for
extraction of the page number from the citation page number match. This
seems to be better, but I'm not convinced that more bugs are lurking here.

But the regexp didn't solve the fundamental problem: non-breaking hyphens.
I added a regexp to find the non-breaking hyphen elements in the raw input,
replacing them with simple ASCII '-' characters.

This seems to have solved the problem for now.

All tests passing.