added citationYearAlt regexp
The last commit still wasn't handling malformed years correctly in citations.

In particular, a number like "161180" was being interpreted as the year 1611.

I got the `citationYear` regexp to handle this along with the other test cases
which I have collected, except for the case when a year comes at the very end
of a line with no punctuation.

I had to add `citationYearAlt` to handle this
case. It accepts one optional punctuation character for a year at the end of
a line. If a citation line matches this, the year at the end of the line is
taken to be the citation year. Otherwise `citationYear` is used to find the
citation year.

Note that this could cause problems if the user tries to put 4 digit page
numbers after the publication year.

I added a couple of test cases to the `test_example_citations.docx` file.

All tests passing.